chore: move operator sign up related code to sign up module
