+++ packages/apps/human-app/frontend/src/modules/signup/operator/components/add-keys/index.ts
@@ -1,0 +1,2 @@
+export * from './edit-existing-keys-form';
+export * from './pending-keys-form';
