+++ packages/apps/human-app/frontend/src/modules/signup/operator/components/add-stake/buttons.tsx
