+++ packages/apps/human-app/frontend/src/modules/signup/operator/components/add-stake/index.ts
@@ -1,0 +1,2 @@
+export * from './buttons';
+export * from './stake-form';
