+++ packages/apps/human-app/frontend/src/modules/signup/operator/hooks/use-get-stacked-amount.ts
