+++ packages/apps/human-app/frontend/src/modules/signup/operator/index.ts
@@ -1,0 +1,1 @@
+export * from './views';
