+++ packages/apps/human-app/frontend/src/modules/signup/operator/views/add-stake.page.tsx
@@ -8,17 +8,16 @@ import {
   PageCard,
 } from '@/shared/components/ui/page-card';
 import { getErrorMessageForError } from '@/shared/errors';
-import { Buttons } from '@/modules/operator/components/sign-up/add-stake/buttons';
-import { StakeForm } from '@/modules/operator/components/sign-up/add-stake/stake-form';
 import { Alert } from '@/shared/components/ui/alert';
-import {
-  stakedAmountFormatter,
-  useGetStakedAmount,
-} from '@/modules/operator/hooks/use-get-stacked-amount';
-import { useAddStakeMutationState } from '@/modules/operator/hooks/use-add-stake-mutation-state';
-import { useHMTokenDecimals } from '@/modules/operator/hooks/use-human-token-decimals';
 import { useColorMode } from '@/shared/contexts/color-mode';
 import { onlyDarkModeColor } from '@/shared/styles/dark-color-palette';
+import {
+  stakedAmountFormatter,
+  useAddStakeMutationState,
+  useGetStakedAmount,
+  useHMTokenDecimals,
+} from '../hooks';
+import { StakeForm, Buttons } from '../components/add-stake';
 
 export function AddStakeOperatorPage() {
   const { colorPalette, isDarkMode } = useColorMode();
