+++ packages/apps/human-app/frontend/src/modules/signup/operator/views/index.ts
@@ -1,0 +1,5 @@
+export * from './add-keys.page';
+export * from './add-stake.page';
+export * from './connect-wallet.page';
+export * from './set-up-operator.page';
+export * from './edit-existing-keys-success.page';
