+++ packages/apps/human-app/frontend/src/modules/signup/operator/views/set-up-operator.page.tsx
