+++ packages/apps/human-app/frontend/src/router/routes.tsx
@@ -14,11 +14,6 @@ import {
   WorkHeaderIcon,
 } from '@/shared/components/ui/icons';
 import type { PageHeaderProps } from '@/shared/components/layout/protected/page-header';
-import { SetUpOperatorPage } from '@/modules/operator/views/sign-up/set-up-operator.page';
-import { EditExistingKeysSuccessPage } from '@/modules/operator/views/sign-up/edit-existing-keys-success.page';
-import { AddKeysOperatorPage } from '@/modules/operator/views/sign-up/add-keys.page';
-import { AddStakeOperatorPage } from '@/modules/operator/views/sign-up/add-stake.page';
-import { ConnectWalletOperatorPage } from '@/modules/operator/views/sign-up/connect-wallet.page';
 import { Playground } from '@/modules/playground/views/playground.page';
 import {
   HcaptchaLabelingPage,
@@ -35,6 +30,13 @@ import { WorkerProfilePage } from '@/modules/worker/profile';
 import { SignUpWorkerPage } from '@/modules/signup/worker';
 import { OperatorProfilePage } from '@/modules/operator/profile';
 import { HomePage } from '@/modules/homepage';
+import {
+  AddKeysOperatorPage,
+  AddStakeOperatorPage,
+  ConnectWalletOperatorPage,
+  EditExistingKeysSuccessPage,
+  SetUpOperatorPage,
+} from '@/modules/signup/operator';
 
 export const unprotectedRoutes: RouteProps[] = [
   {
